Add 'try again' to compile.sh, rename web_handler
Adds 'try again' functionality to compile.sh in response to frequent
build/install errors brought about by the use of Bash for Windows.

<enter> must be pressed for compile.sh to run through build/install
again. This is to prevent looping in cases where build/install fails
due to errors with the code (rather than with Bash for Windows).

Renames web_handler.go to generate_url.go, as there will likely be
another file which actually handles http requests that will be named
web_handler in the future.

generate_url is a more descriptive name of what the code in that file
does.