Remove StatusCode from download, fix pointers
download.StatusCode was originally added as a shorthand for
download.Response.StatusCode, but since it is not going to be used
often, it will be removed for simplicity. Change will be reverted if
StatusCode sees more use.

download now contains pointers to the overlayUrl used to generate it,
as well as the http.Response created from overlayUrl.Url, instead of
values.