Add csv, image relevant functionality
Add new csv.go, includes two (out of three) public methods
AppendCSV and GetLastEntry that can be used to read and modify the
record.csv file easily.

Add public function GetImage to download.go which returns an
image.Image type which can then be manipulated by relevant methods
in image.go; this saves the unnecessary step of saving the image
to file (and reloading it).

Renamed folderUtils.go to osUtils.go to better describe its scope.

Add image.go. Includes public methods to manipulate overlay, map,
etc.

Add res/mask.png, which is used in image.go -> AlphaPercentMask()
to calculate percentage of (overlay covering land area in map /
land area in map) as opposed to the old AlphaPercent, which
calculates percentage of (overlay covering map / map).

NOTE: The generation of overlay + map with AddMap is not confirmed
to be working reliably. This issue is not consistently reproducible.
See: http://stackoverflow.com/questions/43959864/image-glitch-using-golangs-drawmask-with-draw-over