Add twitter.go (currently only decodes json)
keys.json is a file in the working directory that will store API Keys
and other sensitive information.

keys.json is decoded in twitter.go; additional functionality in this
file to POST to the twitter API is a to-do

Also moved panic(err) line down one in gif.go for consistency.